+++ buildSrc/src/main/kotlin/dev/greenhouseteam/bovinesandbuttercups/gradle/Properties.kt
@@ -3,6 +3,7 @@ package dev.greenhouseteam.bovinesandbuttercups.gradle
 object Properties {
     const val GROUP = "dev.greenhouseteam"
     const val MOD_NAME = "Bovines and Buttercups"
+    const val ARCHIVES_NAME = "bovines-and-buttercups"
     const val MOD_ID = "bovinesandbuttercups"
     const val MOD_AUTHOR = "Greenhouse Team"
     val MOD_CONTRIBUTORS = listOf("MerchantPug", "SpiderKolo", "Omniv")